pyproject.toml
[project]
name = "notebook-pdf"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mads-Peter", email = "machri@phys.au.dk" }
]
requires-python = ">=3.12"
dependencies = [
    "ipython",
    "ipywidgets>=8.1.8",
    "nbformat>=5.10.4",
    "pyyaml>=6.0.3",
    "requests",
    "werkzeug>=3.1.6",
]

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pip>=26.0.1",
]
